Simple Webpage Framework: Create and add
We'd like a basic website framework to make webpages look pretty.

I'm trying not to rely on a specific framework, but I'd still like
a basic template engine to separate HTML from PHP.

I found this framework on a simple google search:
https://codeshack.io/lightweight-template-engine-php/
and it does basically what I want.

I changed a couple things to make it work in `src/templates` and to
not requre `.html` when trying to build the template.

This also added an autoload, .gitignore and moved app.php into api/index.php
so we could get the "api response" from "/api".