Framework Setup: Add Bootstrap and simple styles
We'd like to add some simple styles to our site.

Bootstrap is a pretty easy way of doing that and I'm familiar with
it so we'll add that code and do some basic css to show it working.

This also adds some JS that we aren't using but it's a set of libraries
I'll probably use at some point so nbd.

Mostly followed stuff at https://getbootstrap.com/