JQuery: Update and add test file
There's an update to JQuery we can do so we might as well.

Also adds a simple js file in another file and loads it b/c we're
probably gonna wanna add JS and we don't want it all on the base
page.

Also adds the "defer" tag so we run it after all our framework stuff
loads.